Add files via upload
This is what I currently have completed. Works for a single main account to store as many usernames and passwords as desired and has built-in search functionality for a specific user name and password and can display all usernames and passwords stored on the main account. Also has a built-in password generator. What is missing: Multiple main account functionality, documentation/comments for the functions of the code, separation of functions into multiple class and header files to make the code easier to read.